refactor: 테스트 패키지 분리 및 @Transactional 어노테이션 추가
