+++ src/test/java/kitchenpos/application/menu/MenuServiceTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.application;
+package kitchenpos.application.menu;
 
 import com.sun.tools.javac.util.List;
 import kitchenpos.menu.domain.Menu;
