+++ src/test/java/kitchenpos/application/product/ProductServiceTest.java
@@ -1,20 +1,22 @@
-package kitchenpos.application;
+package kitchenpos.application.product;
 
-import kitchenpos.product.domain.Product;
-import kitchenpos.product.application.dto.ProductRequest;
 import kitchenpos.product.application.ProductService;
+import kitchenpos.product.application.dto.ProductRequest;
+import kitchenpos.product.domain.Product;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.jdbc.Sql;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.math.BigDecimal;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+@Transactional
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.NONE)
 @Sql({"/h2-truncate.sql"})
 class ProductServiceTest {
